+++ tools/ddsrouter_yaml_validator/package.xml
@@ -2,9 +2,11 @@
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
   <name>ddsrouter_yaml_validator</name>
-  <version>0.0.0</version>
+  <version>0.3.0</version>
   <description>Tool used for validating DDS-Router configuration files</description>
-  <maintainer email="juanlopez@eprosima.com">eprosima</maintainer>
+  <maintainer email="raul@eprosima.com">Raúl Sánchez-Mateos</maintainer>
+  <maintainer email="javierparis@eprosima.com">Javier París</maintainer>
+  <maintainer email="juanlopez@eprosima.com">Juan López</maintainer>
   <license>Apache License, Version 2.0</license>
 
   <test_depend>python3-pytest</test_depend>
