+++ tools/ddsrouter_yaml_validator/tests/invalid_configuration_files/address_no_ip_nor_domain.yaml
