+++ tools/ddsrouter_yaml_validator/tests/invalid_configuration_files/builtin_topic_no_type.yaml
