+++ tools/ddsrouter_yaml_validator/tests/invalid_configuration_files/ds_participant_no_discovery_server_guid.yaml
