+++ tools/ddsrouter_yaml_validator/tests/invalid_configuration_files/ds_participant_no_listening_nor_connection_addresses.yaml
