+++ tools/ddsrouter_yaml_validator/tests/invalid_configuration_files/filter_topic_no_name.yaml
