+++ tools/ddsrouter_yaml_validator/tests/invalid_configuration_files/no_participant_kind.yaml
