+++ tools/ddsrouter_yaml_validator/tests/invalid_configuration_files/tls_ca_no_private_key_provided_cert.yaml
